extracted card test to separate file. added mox library so that we can use mocks in unit tests
